// packages/sdk/src/api/types.gen.ts
export type NfdProperties = {
  /**
   * Internal properties
   */
  internal?: {
    [key: string]: string
  }
  /**
   * User properties
   */
  userDefined?: {
    [key: string]: string
  }
  /**
   * Verified properties
   */
  verified?: {
    [key: string]: string
  }
}
// ...
export type NfdRecord = {
  /**
   * NFD Application ID
   */
  appID?: number
  /**
   * NFD ASA ID
   */
  asaID?: number
  /**
   * Whether the verified Avatar set in this NFD is newer (arc19) then is set into the NFD. This will only be present on direct NFD fetch and if true
   */
  avatarOutdated?: boolean
  /**
   * Verified Algorand addresses for this NFD
   */
  caAlgo?: Array<string>
  /**
   * Cache-Control header
   */
  'cache-control'?: string
  /**
   * Category of NFD
   */
  category?: 'curated' | 'premium' | 'common'
  /**
   * Round this data was last fetched from
   */
  currentAsOfBlock?: number
  /**
   * An Algorand Account address
   */
  depositAccount?: string
  /**
   * ETag
   */
  etag?: string
  expired?: boolean
  /**
   * Not returned, used in tagging for response to indicate if-none-match etag matched
   */
  'match-check'?: string
  /**
   * Tags set by the system for tracking/analytics
   */
  metaTags?: Array<string>
  name: string
  /**
   * An Algorand Account address
   */
  nfdAccount?: string
  /**
   * An Algorand Account address
   */
  owner?: string
  /**
   * NFD Application ID of Parent if this is a segment
   */
  parentAppID?: number
  properties?: NfdProperties
  /**
   * An Algorand Account address
   */
  reservedFor?: string
  /**
   * Sale type of NFD
   */
  saleType?: 'auction' | 'buyItNow'
  /**
   * amount NFD is being sold for (microAlgos)
   */
  sellAmount?: number
  /**
   * An Algorand Account address
   */
  seller?: string
  /**
   * An Algorand Account address
   */
  sigNameAddress?: string
  /**
   * State of NFD
   */
  state?: 'available' | 'minting' | 'reserved' | 'forSale' | 'owned' | 'expired'
  /**
   * Tags assigned to this NFD
   */
  tags?: Array<string>
  timeChanged?: string
  timeCreated?: string
  timeExpires?: string
  timePurchased?: string
  /**
   * Unverified (non-algo) Crypto addresses for this NFD
   */
  unverifiedCa?: {
    [key: string]: Array<string>
  }
  /**
   * Unverified Algorand addresses for this NFD
   */
  unverifiedCaAlgo?: Array<string>
}